refactor(aceditor): better javascript for actions builder and flying button
