+++ tools/aceditor/lang/aceditor_en.inc.php
@@ -39,8 +39,7 @@ $GLOBALS['translations'] = array_merge($GLOBALS['translations'], array(
     'ACEDITOR_LINK_PAGE_NAME'           => 'YesWiki page name or URL',
     'ACEDITOR_LINK_HINT_NEW_PAGE_NAME'  => 'If the page does not exists, it will be created.',
     'ACEDITOR_LINK_TEXT'                => 'Text displayed',
-    'ACEDITOR_LINK_OPEN_IN_CURRENT_TAB' => 'Link opens in current tab (old syntax)',
-    'ACEDITOR_LINK_OPEN_IN_CURRENT_TAB_ACTION_SYNTAX'=> 'Link opens in current tab',
+    'ACEDITOR_LINK_OPEN_IN_CURRENT_TAB' => 'Link opens in current tab',
     'ACEDITOR_LINK_OPEN_IN_NEW_TAB'     => 'Link opens in new tab',
     'ACEDITOR_LINK_OPEN_IN_MODAL'       => 'Link opens in modal box',
     'ACEDITOR_LINK_CANCEL'              => 'Cancel',
