+++ tools/aceditor/lang/aceditor_fr.inc.php
@@ -38,8 +38,7 @@ $GLOBALS['translations'] = array_merge($GLOBALS['translations'], array(
     'ACEDITOR_LINK_PAGE_NAME'          => 'Nom de la page ou URL externe',
     'ACEDITOR_LINK_HINT_NEW_PAGE_NAME' => "Si la page n'existe pas, elle sera créée.",
     'ACEDITOR_LINK_TEXT'               => 'Texte affiché',
-    'ACEDITOR_LINK_OPEN_IN_CURRENT_TAB'=> 'Le lien s\'ouvre dans l\'onglet courant (ancienne syntaxe)',
-    'ACEDITOR_LINK_OPEN_IN_CURRENT_TAB_ACTION_SYNTAX'=> 'Le lien s\'ouvre dans l\'onglet courant',
+    'ACEDITOR_LINK_OPEN_IN_CURRENT_TAB'=> 'Le lien s\'ouvre dans l\'onglet courant',
     'ACEDITOR_LINK_OPEN_IN_NEW_TAB'    => 'Le lien s\'ouvre dans un nouvel onglet',
     'ACEDITOR_LINK_OPEN_IN_MODAL'      => 'Le lien s\'ouvre dans une fenêtre modale',
     'ACEDITOR_LINK_CANCEL'             => 'Annuler',
