+++ tools/aceditor/presentation/javascripts/aceditor.js
@@ -6,68 +6,51 @@ import setupAceditorKeyBindings from './aceditor-key-bindings.js'
 import openModal from './aceditor-toolbar-remote-modal.js'
 import LinkModal from './link-modal.js'
 import FileUploadModal from '../../../attach/presentation/javascripts/file-upload-modal.js'
+import AceWrapper from './ace-wrapper.js'
+import ActionsBuilder from './actions-builder.js'
+import FlyingEditButton from './flying-edit-button.js'
 
 class Aceditor {
-  ace = null
+  editor
   linkModal
+  fileUplodModal
 
   constructor($container) {
     this.$container = $container
     this.initialize()
   }
 
-  get $textarea() {
-    return this.$container.find('.aceditor-textarea')
-  }
-
-  get $aceContainer() {
-    return this.$container.find('.ace-container')
-  }
-
-  get $toolbar() {
-    return this.$container.find('.aceditor-toolbar')
-  }
+  get $textarea() { return this.$container.find('.aceditor-textarea') }
+  get $aceContainer() { return this.$container.find('.ace-container') }
+  get $aceBody() { return this.$container.find('.ace-body') }
+  get $toolbar() { return this.$container.find('.aceditor-toolbar') }
 
   initialize() {
-    // Where to find the 'mode-XXXX' files
-    ace.config.set('basePath', 'tools/aceditor/presentation/javascripts')
-
-    this.ace = ace.edit(this.$container.find('.ace-body')[0], {
-      printMargin: false,
-      mode: 'ace/mode/yeswiki',
-      showGutter: true,
-      wrap: 'free',
-      maxLines: Infinity,
-      minLines: this.$textarea.attr('rows'),
-      showFoldWidgets: false,
-      fontSize: '18px',
-      useSoftTabs: false,
-      tabSize: 3,
-      fontFamily: 'monospace',
-      highlightActiveLine: true
-    })
+    // Init Components
+    this.editor = new AceWrapper(this.$aceBody[0], { rows: this.$textarea.attr('rows') })
+    this.linkModal = new LinkModal()
+    this.fileUplodModal = new FileUploadModal()
+    this.actionsBuilder = new ActionsBuilder()
+    this.flyingButton = new FlyingEditButton(this.$container)
 
     // Sync textarea and editor
-    this.ace.session.setValue(this.$textarea.val())
-    this.ace.session.on('change', () => {
-      this.$textarea.val(this.ace.session.getValue())
-    })
-
-    setupAceditorKeyBindings(this.$aceContainer, this.$toolbar)
-
-    // Enable alert popup when leaving the page
-    this.ace.on('change', () => {
+    this.editor.setValue(this.$textarea.val())
+    this.editor.on('change', () => {
+      this.$textarea.val(this.editor.getValue())
+      // Enable alert popup when leaving the page
       if (typeof showPopup !== 'undefined') { showPopup = 1 }
     })
 
-    this.linkModal = new LinkModal()
-    this.fileUplodModal = new FileUploadModal()
+    setupAceditorKeyBindings(this.$aceContainer, this.$toolbar)
+    this.initToolbar()
+    this.initEditionHelpers()
+  }
 
+  initToolbar() {
     this.fileUplodModal.initButton(
       this.$toolbar.find('.attach-file-uploader'),
-      (result) => { this.replaceSelectionBy(result) }
+      (result) => { this.editor.replaceSelectionBy(result) }
     )
-
     this.$toolbar.find('.aceditor-btn').on('click', (e) => {
       const $btn = $(e.currentTarget)
 
@@ -77,22 +60,57 @@ class Aceditor {
       } else if ($btn.hasClass('aceditor-btn-link')) {
         // Link Button
         this.linkModal.open({
-          text: this.ace.getSelectedText(),
-          onComplete: (result) => { this.replaceSelectionBy(result) }
+          text: this.editor.getSelectedText(),
+          onComplete: (result) => { this.editor.replaceSelectionBy(result) }
         })
       } else {
         // Other Buttons
-        this.surroundSelectionWith($btn.data('lft'), $btn.data('rgt'))
+        this.editor.surroundSelectionWith($btn.data('lft'), $btn.data('rgt'))
       }
+    })
+    this.$toolbar.find('.open-actions-builder-btn').click((event) => {
+      this.actionsBuilder.open({ groupName: $(event.target).data('group-name') })
+    })
+    this.$toolbar.find('.open-existing-action').click(() => {
+      this.actionsBuilder.open({ action: this.editor.currentGroupTextwithoutMarkup })
     })
   }
 
-  surroundSelectionWith(left = '', right = '') {
-    this.ace.session.replace(this.ace.getSelectionRange(), left + this.ace.getSelectedText() + right)
-  }
+  initEditionHelpers() {
+    this.editor.onCursorChange((cursor) => {
+      // Reset
+      this.$toolbar.find('.component-action-list').removeClass('only-edit')
+      this.flyingButton.hide()
 
-  replaceSelectionBy(replacement) {
-    this.ace.session.replace(this.ace.getSelectionRange(), replacement)
+      // wait for the full group to be written
+      if (!cursor.groupType || !cursor.groupEnd) return
+
+      switch (cursor.groupType) {
+        case 'yw-action': {
+          const [actionName] = cursor.groupTextWithoutMarkup.split(' ')
+          if (this.actionsBuilder.allAvailableActions.includes(actionName)) {
+            this.$toolbar.find('.component-action-list').addClass('only-edit')
+            this.flyingButton.show().onClick(() => {
+              this.actionsBuilder.open(this.editor, { action: cursor.groupTextWithoutMarkup })
+            })
+          }
+          break
+        }
+        case 'yw-link': {
+          const [link, text] = cursor.groupTextWithoutMarkup.split(' ')
+          this.flyingButton.show().onClick(() => {
+            this.linkModal.open({
+              link,
+              text,
+              onComplete: (result) => { this.editor.replaceCurrentGroupBy(result) }
+            })
+          })
+          break
+        }
+        default:
+          break
+      }
+    })
   }
 }
