+++ tools/aceditor/presentation/javascripts/flying-edit-button.js
@@ -1,0 +1,32 @@
+export default class {
+  $aceditor
+
+  constructor($aceditorContainer) {
+    this.$aceditor = $aceditorContainer
+  }
+
+  get $flyingButton() {
+    return this.$aceditor.find('.flying-edit-button')
+  }
+
+  show() {
+    // wait for editor to change cursor
+    setTimeout(() => {
+      this.$flyingButton.addClass('active')
+      const top = this.$aceditor.find('.ace_gutter-active-line').offset().top
+        - this.$aceditor.find('.ace-container').offset().top
+        + this.$aceditor.find('.aceditor-toolbar').height()
+      this.$flyingButton.css('top', `${top}px`)
+    }, 100)
+    return this
+  }
+
+  hide() {
+    this.$flyingButton.removeClass('active')
+  }
+
+  onClick(callback) {
+    this.$flyingButton.off('click')
+    this.$flyingButton.on('click', callback)
+  }
+}
