+++ tools/aceditor/templates/aceditor.twig
@@ -8,6 +8,12 @@
     <pre class="ace-body"></pre>
   </div>
 
+  <div class="flying-edit-button">
+    <a class="open-actions-builder-btn btn btn-primary btn-icon">
+      <i class="fa fa-pencil-alt"></i>
+    </a>
+  </div>
+
   <textarea id="{{ arguments.name }}" name="{{ arguments.name }}" class="aceditor-textarea">
     {{- arguments.value|raw -}}
   </textarea>
