+++ tools/aceditor/templates/link-modal.twig
@@ -22,8 +22,7 @@
 						</div>
 					</div>
 					<select class="form-control link-options" style="margin: 1.5rem 0 2rem 0;">
-						<option value="link">{{ _t('ACEDITOR_LINK_OPEN_IN_CURRENT_TAB') }}</option>
-						<option value="button" selected>{{ _t('ACEDITOR_LINK_OPEN_IN_CURRENT_TAB_ACTION_SYNTAX') }}</option>
+						<option value="link" selected>{{ _t('ACEDITOR_LINK_OPEN_IN_CURRENT_TAB') }}</option>
 						<option value="ext">{{ _t('ACEDITOR_LINK_OPEN_IN_NEW_TAB') }}</option>
 						<option value="modal">{{ _t('ACEDITOR_LINK_OPEN_IN_MODAL') }}</option>
 					</select>
